+++ foundation_headless/tests.lua
@@ -1,3 +1,4 @@
 local mod = assert(foundation_headless)
 
+mod:require("tests/inv_ref_test.lua")
 mod:require("tests/item_stack_test.lua")